okay la fin

+++ ft_atoi.c
@@ -6,17 +6,17 @@
 /*   By: yschecro <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/19 21:40:15 by yschecro          #+#    #+#             */
-/*   Updated: 2022/07/10 15:50:27 by yschecro         ###   ########.fr       */
+/*   Updated: 2022/07/10 21:56:34 by yschecro         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "push_swap.h"
 
-int	ft_atoi(char *str)
+long	ft_atoi(char *str)
 {
-	int	i;
-	int	out;
-	int	signe;
+	int		i;
+	long	out;
+	int		signe;
 
 	i = 0;
 	signe = 1;
